+++ backend/src/main/java/com/laioffer/flagcamp/backend/entity/User.java
@@ -19,7 +19,7 @@ public record User(
         
         @NotNull
         @Column("phonenum")
-        Integer phoneNum,
+        Long phoneNum,
 
         @NotNull
         String address,
